fix(node): :zap: make `node` port browser-compatible
