+++ src/node/src/cli.ts
@@ -1,12 +1,9 @@
 #!/usr/bin/env node
 
-import {
-	convertFile,
-	formatHelpMessage,
-	type GhmdMode,
-	type GhmdTheme,
-	strings,
-} from './index.js'
+import { readFileSync, writeFileSync } from 'node:fs'
+import { strings } from './generated/shared.js'
+import type { GhmdMode, GhmdTheme } from './index.js'
+import { convertMarkdownToHtml, formatHelpMessage } from './index.js'
 
 interface ParsedCliArguments {
 	files: string[]
@@ -36,6 +33,23 @@ function parseCliArguments(args: string[]): ParsedCliArguments {
 	}
 
 	return { files, help, theme, embedCss, mode }
+}
+
+async function convertFile(
+	inputPath: string,
+	options: { theme: GhmdTheme; embedCss: boolean; mode: GhmdMode },
+) {
+	if (inputPath.endsWith('.html')) throw new Error(strings.errors.htmlExtension)
+
+	const markdown = readFileSync(inputPath, 'utf-8')
+	const html = await convertMarkdownToHtml(markdown, {
+		...options,
+		token: process.env.GITHUB_TOKEN,
+	})
+	const outputPath = `${inputPath.split('.').slice(0, -1).join('.') || inputPath}.html`
+
+	writeFileSync(outputPath, html)
+	return outputPath
 }
 
 async function main() {
